minor refactor: move classes to packages
