+++ src/main/scala/com/amazon/deequ/dqdl/EvaluateDataQuality.scala
@@ -16,6 +16,9 @@
 
 package com.amazon.deequ.dqdl
 
+import com.amazon.deequ.dqdl.execution.DeequCheckExecutor
+import com.amazon.deequ.dqdl.translation.{DQDLRuleTranslator, DeequOutcomeTranslator}
+import com.amazon.deequ.dqdl.util.DefaultDQDLParser
 import org.apache.spark.sql.DataFrame
 import software.amazon.glue.dqdl.model.DQRuleset
